requirements.txt
scikit-learn==1.6.1
scipy==1.16.1
numpy==2.0.2
pandas==2.2.3
sklearn-compat==0.1.3
Flask==3.1.0